Update to booking inspector:
* Added review mode and functionality to BookingEditor.svelte

* Removed old BookingInspector.svelte

* Renamed instances of "inspect" to "review"

* Added translations message for admin/[id] page title

* Moved StatusComponent.svelte to reflect its new usage
