+++ src/routes/(app)/booking/BookingEditor.svelte
@@ -3,15 +3,21 @@
   import type { BookingSchema } from "./schema";
   import { superForm } from "$lib/utils/client/superForms";
   import * as m from "$paraglide/messages";
-  import type { Bookable } from "@prisma/client";
+  import type { Bookable, BookingRequest } from "@prisma/client";
+  import StatusComponent from "./StatusComponent.svelte";
 
+  type BookingRequestWithBookables = BookingRequest & { bookables: Bookable[] };
   export let data: {
     form: SuperValidated<Infer<BookingSchema>>;
     bookables: Bookable[];
+    booking?: BookingRequestWithBookables;
+    allBookingRequests?: BookingRequestWithBookables[];
   };
 
+  $: bookingRequest = data.booking;
+
   const { form, errors, enhance, constraints } = superForm(data.form);
-  export let mode: "create" | "edit" = "create";
+  export let mode: "create" | "edit" | "review" = "create";
 
   let start = $form.start;
   let end = $form.end;
@@ -60,6 +66,23 @@
 </script>
 
 <form method="POST" use:enhance class="form-control mx-auto max-w-5xl gap-4">
+  {#if mode === "review"}
+    <div class="flex flex-col gap-5">
+      <div class="w-fit">
+        <a class="btn" href="/booking/admin">
+          <span class="i-mdi-arrow-expand-left" />
+          {m.booking_goBack()}
+        </a>
+      </div>
+      {#if bookingRequest && data.allBookingRequests}
+        <StatusComponent
+          bind:bookingRequest
+          bind:bookingRequests={data.allBookingRequests}
+          mode="review"
+        />
+      {/if}
+    </div>
+  {/if}
   <fieldset
     class="input-bordered grid grid-cols-[repeat(auto-fit,minmax(300px,1fr))] rounded-xl border px-6 py-2"
     class:border-error={$errors.bookables?._errors ?? 0 > 0}
@@ -73,6 +96,7 @@
           name="bookables"
           value={bookable.id}
           bind:group={$form.bookables}
+          disabled={mode === "review"}
         />
         <span class="label-text">{bookable.name}</span>
       </label>
@@ -89,6 +113,7 @@
       bind:value={start}
       on:change={handleStartChange}
       {...$constraints.start}
+      disabled={mode === "review"}
     />
   </label>
 
@@ -103,6 +128,7 @@
       bind:value={end}
       on:change={handleEndChange}
       {...$constraints.end}
+      disabled={mode === "review"}
     />
   </label>
 
@@ -114,15 +140,42 @@
       class="input input-bordered w-full"
       bind:value={$form.name}
       {...$constraints.name}
+      disabled={mode === "review"}
     />
   </label>
 
-  <div class="flex *:flex-1">
-    <a class="btn" href="/booking">{m.booking_goBack()}</a>
-    {#if mode === "edit"}
-      <button class="btn btn-primary">{m.save()}</button>
-    {:else if mode === "create"}
-      <button class="btn btn-primary">{m.booking_create()}</button>
-    {/if}
-  </div>
+  {#if mode === "review" && bookingRequest}
+    <div class="flex *:flex-1">
+      <input hidden name="id" type="text" bind:value={bookingRequest.id} />
+      <button
+        formaction="?/accept"
+        class="btn btn-outline btn-success"
+        class:btn-disabled={bookingRequest.status === "ACCEPTED"}
+        aria-label={m.booking_accept()}
+      >
+        {m.booking_accept()}
+        <span class="i-mdi-check" />
+      </button>
+      <button
+        formaction="?/reject"
+        class="btn btn-outline btn-error"
+        class:btn-disabled={bookingRequest.status === "DENIED"}
+        aria-label={m.booking_deny()}
+      >
+        {m.booking_deny()}
+        <span class="i-mdi-close" />
+      </button>
+    </div>
+  {/if}
+
+  {#if mode !== "review"}
+    <div class="flex *:flex-1">
+      <a class="btn" href="/booking">{m.booking_goBack()}</a>
+      {#if mode === "edit"}
+        <button class="btn btn-primary">{m.save()}</button>
+      {:else if mode === "create"}
+        <button class="btn btn-primary">{m.booking_create()}</button>
+      {/if}
+    </div>
+  {/if}
 </form>
