+++ src/routes/(app)/booking/admin/[id]/+page.svelte
@@ -1,14 +1,15 @@
 <script lang="ts">
   import type { PageData } from "./$types";
   import SetPageTitle from "$lib/components/nav/SetPageTitle.svelte";
-  import BookingInspector from "../BookingInspector.svelte";
   import BookingCalendar from "../../BookingCalendar.svelte";
+  import BookingEditor from "../../BookingEditor.svelte";
+  import * as m from "$paraglide/messages";
 
   export let data: PageData;
 </script>
 
-<SetPageTitle title="Booking Admin Test" />
+<SetPageTitle title={m.booking_reviewBooking()} />
 
-<BookingInspector {data} />
+<BookingEditor {data} mode="review" />
 
 <BookingCalendar {...data} class="mt-16" />
