+++ src/translations/en.json
@@ -755,5 +755,6 @@
   "nollning_wikia_literature": "Student literature",
   "nollning_wikia_literature_body": "Facebook groupo where a lot of second-hand course literature is being sold by older students, it's a private group but everyone is allowed to join, just ask a phadder to approve you.",
   "members_errors_tooLargePicture": "Profile picture too big, max size is {size}.",
-  "nollning_events_ticketCTA": "Ticket/Register"
+  "nollning_events_ticketCTA": "Ticket/Register",
+  "booking_reviewBooking": "Review booking"
 }
